Refactor unit tests for trust pages
Introduce field variables to reduce repeated setup code.
