+++ tests/DfE.FindInformationAcademiesTrusts.UnitTests/Pages/Trusts/ContactsModelTests.cs
@@ -1,73 +1,64 @@
 using DfE.FindInformationAcademiesTrusts.Pages.Trusts;
 using Microsoft.AspNetCore.Mvc;
 
-namespace DfE.FindInformationAcademiesTrusts.UnitTests.Pages;
+namespace DfE.FindInformationAcademiesTrusts.UnitTests.Pages.Trusts;
 
 public class ContactsModelTests
 {
+    private readonly Mock<ITrustProvider> _mockTrustProvider;
+    private readonly ContactsModel _sut;
+
+    public ContactsModelTests()
+    {
+        _mockTrustProvider = new Mock<ITrustProvider>();
+        _sut = new ContactsModel(_mockTrustProvider.Object);
+    }
+
     [Fact]
     public async void OnGetAsync_should_fetch_a_trust_by_ukprn()
     {
-        var mockTrustProvider = new Mock<ITrustProvider>();
-        mockTrustProvider.Setup(s => s.GetTrustByUkprnAsync("1234").Result)
+        _mockTrustProvider.Setup(s => s.GetTrustByUkprnAsync("1234").Result)
             .Returns(new Trust("test", "test", "Multi-academy trust"));
-        var sut = new ContactsModel(mockTrustProvider.Object)
-        {
-            Ukprn = "1234"
-        };
-
-        await sut.OnGetAsync();
-        sut.Trust.Should().BeEquivalentTo(new Trust("test", "test", "Multi-academy trust"));
+        _sut.Ukprn = "1234";
+        await _sut.OnGetAsync();
+        _sut.Trust.Should().BeEquivalentTo(new Trust("test", "test", "Multi-academy trust"));
     }
 
     [Fact]
     public async void Ukprn_should_be_empty_string_by_default()
     {
-        var mockTrustProvider = new Mock<ITrustProvider>();
-        var sut = new ContactsModel(mockTrustProvider.Object);
-
-        await sut.OnGetAsync();
-        sut.Ukprn.Should().BeEquivalentTo(string.Empty);
+        await _sut.OnGetAsync();
+        _sut.Ukprn.Should().BeEquivalentTo(string.Empty);
     }
 
     [Fact]
     public void PageName_should_be_Contacts()
     {
-        var mockTrustProvider = new Mock<ITrustProvider>();
-        var sut = new ContactsModel(mockTrustProvider.Object);
-        sut.PageName.Should().Be("Contacts");
+        _sut.PageName.Should().Be("Contacts");
     }
 
     [Fact]
     public void PageSection_should_be_AboutTheTrust()
     {
-        var mockTrustProvider = new Mock<ITrustProvider>();
-        var sut = new ContactsModel(mockTrustProvider.Object);
-        sut.Section.Should().Be("About the trust");
+        _sut.Section.Should().Be("About the trust");
     }
 
     [Fact]
     public async void OnGetAsync_should_return_not_found_result_if_trust_is_not_found()
     {
-        var mockTrustProvider = new Mock<ITrustProvider>();
-        mockTrustProvider.Setup(s => s.GetTrustByUkprnAsync("1111").Result)
+        _mockTrustProvider.Setup(s => s.GetTrustByUkprnAsync("1111").Result)
             .Returns((Trust?)null);
 
-        var sut = new ContactsModel(mockTrustProvider.Object)
-        {
-            Ukprn = "1111"
-        };
-        var result = await sut.OnGetAsync();
+        _sut.Ukprn = "1111";
+
+        var result = await _sut.OnGetAsync();
         result.Should().BeOfType<NotFoundResult>();
     }
 
     [Fact]
     public async void OnGetAsync_should_return_not_found_result_if_Ukprn_is_not_provided()
     {
-        var mockTrustProvider = new Mock<ITrustProvider>();
-
-        var sut = new ContactsModel(mockTrustProvider.Object);
-        var result = await sut.OnGetAsync();
+        var result = await _sut.OnGetAsync();
         result.Should().BeOfType<NotFoundResult>();
     }
 }
